almost finished apis

+++ dev-data/dataStarter.js
@@ -1,11 +1,17 @@
+//////////////////////////////////////////////////////////////////
+// how to use this
+//node dev-data/dataStarter --import Review reviews.json
+//node dev-data/dataStarter --delete Review
+////////////////////////////////////////////////////////////////////
 const fs = require("fs");
 const mongoose = require("mongoose");
 const dotenv = require("dotenv");
 const Tour = require("./../models/tourModel");
 const User = require("./../models/userModel");
+const Review = require("./../models/reviewModel");
 
 dotenv.config({ path: `./env.env` });
-const tours = JSON.parse(
+const docs = JSON.parse(
   fs.readFileSync(
     `${__dirname}/data/${
       process.argv[4] ? process.argv[4] : "tours-simple.json"
@@ -16,8 +22,9 @@ const tours = JSON.parse(
 const importData = async () => {
   try {
     console.log("start uploade");
-    if (process.argv[3] === "Tour") await Tour.create(tours);
-    if (process.argv[3] === "User") await User.create(tours);
+    if (process.argv[3] === "Tour") await Tour.create(docs);
+    else if (process.argv[3] === "User") await User.create(docs);
+    else if (process.argv[3] === "Review") await Review.create(docs);
     console.log("Data successfully loaded!");
   } catch (err) {
     console.log(err);
@@ -27,7 +34,9 @@ const importData = async () => {
 const deleteData = async () => {
   try {
     console.log("start deleting....");
-    await Tour.deleteMany({});
+    if (process.argv[3] === "Tour") await Tour.deleteMany({});
+    else if (process.argv[3] === "User") await User.deleteMany({});
+    else if (process.argv[3] === "Review") await Review.deleteMany({});
     console.log("finish deleting....");
   } catch (err) {
     console.log(err);
